refactoring for multiplatform build
